refactor into separate files, account for null errors, use loglevel
